requirements.txt
Flask==2.3.2
gunicorn==20.1.0
google-analytics-data==0.14.0